requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-core
langchain-openai
langchain-google-genai
langchain-text-splitters
faiss-cpu
sentence-transformers
tiktoken